예약번호 생성 구현 (Todo : 남은 기간별 환불금액 결정, merchan uid값 난수화)

+++ src/main/java/com/yeoyeo/application/dateroom/dto/DateRoomInfoDto.java
@@ -12,7 +12,7 @@ public class DateRoomInfoDto {
     private final LocalDate date;
     private final long roomId;
     private final String roomName;
-    private final long price;
+    private final int price;
     private final long reservationState;
 
     public DateRoomInfoDto(DateRoom entity) {
